Handle leaderboard logic outside of command class
This should (more properly) resolve the issue of page caching. The original cause was because the LeaderboadCommand class is only ever instantiated once, so the page cache Map exists as the same for each instance. I just cleared it on each command run before, but this should make it so that way the map cache uniquely exists for each user in a command prompt.